+++ django/applications/catmaid/static/js/tools/navigator.js
@@ -166,20 +166,21 @@
     var changeSliceDelayedAction = function()
     {
       window.clearTimeout( changeSliceDelayedTimer );
-      self.changeSlice( changeSliceDelayedParam.z );
+      self.changeSlice( changeSliceDelayedParam.z, changeSliceDelayedParam.step );
       changeSliceDelayedParam = null;
       return false;
     };
 
-    this.changeSliceDelayed = function( val )
+    this.changeSliceDelayed = function(val, step)
     {
       if ( changeSliceDelayedTimer ) window.clearTimeout( changeSliceDelayedTimer );
-      changeSliceDelayedParam = { z : val };
+      changeSliceDelayedParam = { z : val, step: step };
       changeSliceDelayedTimer = window.setTimeout( changeSliceDelayedAction, 50 );
     };
 
-    this.changeSlice = function( val )
+    this.changeSlice = function(val, step)
     {
+      val = self.stackViewer.toValidZ(val, step < 0 ? -1 : 1);
       self.stackViewer.moveToPixel( val, self.stackViewer.y, self.stackViewer.x, self.stackViewer.s );
     };
 
@@ -204,7 +205,7 @@
         }
         lastFrameTime = thisFrameTime;
 
-        var zOffset = self.stackViewer.primaryStack.validZDistanceByStep(self.slider_z.val, step);
+        var zOffset = self.stackViewer.validZDistanceByStep(self.slider_z.val, step);
         if (!zOffset) return;
         self.stackViewer.moveToPixel(
             self.slider_z.val + zOffset,