+++ django/applications/catmaid/static/js/widgets/settings.js
@@ -322,6 +322,21 @@
           'prefer_webgl',
           SETTINGS_SCOPE));
 
+      // WebGL tile layers
+      ds.append(wrapSettingsControl(
+          CATMAID.DOM.createCheckboxSetting(
+              "Skip broken sections of extra tile layers by default",
+              CATMAID.StackViewer.Settings[SETTINGS_SCOPE].respect_broken_sections_new_stacks,
+              'Choose whether layers added after the first one should be ' +
+              'respected by default when checking for broken sections during ' +
+              'navigation.',
+              function() {
+                CATMAID.StackViewer.Settings[SETTINGS_SCOPE].respect_broken_sections_new_stacks = this.checked;
+              }),
+          CATMAID.StackViewer.Settings,
+          'respect_broken_sections_new_stacks',
+          SETTINGS_SCOPE));
+
       // Major section step size
       ds.append(wrapSettingsControl(
           CATMAID.DOM.createNumericInputSetting(